Move POS access session construction into a dedicated factory
