+++ Modules/Sources/PointOfSale/Roles/Models/POSAuthError.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public enum POSAuthError: Error, Equatable {
+enum POSAuthError: Error, Equatable {
     case invalidPIN
     case rateLimited(until: Date)
     case permanentlyLocked
