+++ Modules/Sources/PointOfSale/Roles/Models/POSCapability.swift
@@ -1,4 +1,4 @@
-public enum POSCapability: String, CaseIterable, Sendable {
+enum POSCapability: String, CaseIterable, Sendable {
     // Raw values match the WordPress capability identifiers reported for a staff member.
     case viewPOS = "view_pos"
     case viewPOSSettings = "view_pos_settings"
